i think this works

+++ srcs/requirements/wordpress/tools/entrypoint.sh
@@ -2,13 +2,18 @@
 
 cd /var/www/html
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+if [ ! -f /usr/local/bin/wp ]; then
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+fi
 
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
-
-wp core download --allow-root
-wp config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-wp core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+if ! wp core is-installed 2>/dev/null; then
+    wp core download --allow-root --path="/var/www/html"
+    wp config create --dbname=${MYSQL_DATABASE} --dbuser=${MYSQL_USER} --dbpass=$(cat /run/secrets/db_password) --dbhost=mariadb --allow-root
+    wp db create --allow-root 2>/dev/null || echo "Database already exists"
+    wp core install --url=localhost --title="Inception" --admin_user=${WP_ADMIN} --admin_password=$(cat /run/secrets/wp_admin_password) --admin_email=qmennen@student.codam.nl --skip-email --allow-root
+    wp user create --allow-root $WP_AUTHOR $WP_AUTHOR_EMAIL --role=author --user_pass=$(cat /run/secrets/wp_password) --path=/var/www/html
+fi
 
 php-fpm8.2 -F
